virt: replace CPU pinning for migrationCreate
Before the migration is started by libvirt VDSM will first define the VM
on destination host. For that it uses domain XML passed in
VM.migrationCreate call and in it we need to replace existing CPU
pinning with new CPU pinning. Otherwise the define call may fail if the
pCPU indices are out of range.

We already replace pinning in domain XML that we pass to libvirt when we
initiate the migration (libvirt.VIR_MIGRATE_PARAM_DEST_XML parameter).
This XML is migration XML and is different (different content, serves
different purpose) from the XML we pass in VM.migrationCreate. We have
to replace pinning in both separately.

Note that in VM.migrationCreate call we pass two XMLs -- one in `xml`
argument and one in `_srcDomXML`. We replace pinning only in
`_srcDomXML` because that is the one passed to define call on
destination. The choice not to replace pinning in `xml` is arbitrary,
but the rationale may be to avoid doing unnecessary work or to treat
`xml` as initial configuration from engine.

Bug-Url: https://bugzilla.redhat.com/2099321
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>