lvmcmd: remove LVMRunner class
LVMRunner does not have instance members,
so it is a purely static class, which is not
common practice in Python to have.
While static (or class) methods and variables are
accepted, full static class usually means a design flaw.

In this case the LVMRunner class can be removed,
its class methods converted to normal functions,
and the static member moved to a global variable.

Signed-off-by: Albert Esteve <aesteve@redhat.com>