lvmcmd: homogenize command type for lvs
lv commands declared tuples and lists to pass to
commands.start. It is fine as any iterable should
work. However, it is better to have all functions
use the same type and style, and the docstring in
start documents the args as a list, so we change
all types to list.

Signed-off-by: Albert Esteve <aesteve@redhat.com>