hooks: Add log-firmware
This is a combination of the interface of log-console and the
functionality of qemucmdline, hard-coded to log the firmware output to a
log file. qemucmdline can't be used directly for HostedEngine, because
we want a separate log file per VM, and qemucmdline does not have
internally functionality to allow doing that, relying instead on the
engine to supply correct params somehow. For HostedEngine, we can't rely
on the engine, as it does not exist at this point.

Change-Id: I5e2324009daed5136ab7884226cce0feee2e0c96
Signed-off-by: Yedidyah Bar David <didi@redhat.com>