multipath: make isconfigured fail if ufn enabled
Vdsm does not support 'user_friendly_names' in the
multipath configuration, which is currently configured
accordingly by the host. However, this configuration
can be overriden, potentially causing lvm corruption,
and Vdsm does not check for this configuration at all.

This could be checked in 'vdsm-tool isconfigured' call
to prevent Vdsm to start after an upgrade if the
configuration is not correct.

If the attribute is enabled, fail in the isconfigured
call from the tool.multipath module and print the
offending section and attributes. This will prevent
Vdsm to start and will point the user to the right
fix. Example:

$ vdsm-tool is-configured --module multipath
WARNING: Invalid configuration: 'user_friendly_names' is
enabled in multipath configuration:
  section1 {
    key1 value1
    user_friendly_names yes
    key2 value2
  }
  section2 {
    user_friendly_names yes
  }
This configuration is not supported and may lead to storage domain corruption.

Bug-Url: https://bugzilla.redhat.com/1793207
Signed-off-by: Albert Esteve <aesteve@redhat.com>