betterAsyncore: Always close on handle_close
Asyncore calls handle_close() when the peer closed the other side of the
socket. We must close the socket which removes it from asyncore map. If we
don't do that, asyncore may enter a busy loop trying to read or wrtie from a
closed socket, consuming 100% cpu for many seeconds.

Bug-Url: https://bugzilla.redhat.com/2102678
Signed-off-by: Nir Soffer <nsoffer@redhat.com>