betterAsyncore: Fix busy loop after socket is closed
After a peer closed the other size of the socket, socket.recv() returns
b"" (empty bytes), but we wrongly checked for "" (empty str). Because we
did not close the socket, the reactor entered a busy loop calling
handle_read(). The busy loop probably ends when the reactor tries to
send a heartbeat and detect that the socket was closed.

Fixed by checking for b"" and returning b"" to the caller instead of "".

Before this fix, a single vdsm-client run caused the `JsonRpc
(StompReactor))` to consume about 100% cpu for 22 seconds:

    vdsm-client Host getStorageRepoStats

With this fix, we can run vdsm-client in a loop:

    for n in $(seq 100); do
        vdsm-client Host getStorageRepoStats > /dev/null;
    done

And the `JsonRpc (StompReactor))` thread consume less then 1% CPU.

Bug-Url: https://bugzilla.redhat.com/2102678
Signed-off-by: Nir Soffer <nsoffer@redhat.com>