virt: Don’t retrieve VM external data until the VM is started
VM external data, if not provided by Engine, may not be available
initially and may be created only once the VM is started by libvirt.
If its retrieval is attempted before the data is available, an error
message is produced in the logs.  This error message is harmless but
incorrect and it may confuse users or prevent distinguishing real
errors from fake errors.

Let’s not retrieve external data in all the VM initial states, not
only MIGRATION_DESTINATION, to prevent the external data retrieval
error.