blockVolume: check parent tag and meta match
blockVolume getParent method uses the volume tag
PU_<id> to obtain the parent. However, this tag
could be outdated. For example, when the host
syncs a volume chain after removing a volume,
it only updates the metadata. Afterwards,
the SPM will update the volume tag once the
volume is deleted.

To ensure we do not return outdated parent UUID,
we need to ensure that the parent UUID obtained
from the parent tag, and the UUID in the volume
metadata match. If they do not, we try to reload
the volume tag first. If they still do not match
after that, log a warning and use the metadata parent.

Bug-Url: https://bugzilla.redhat.com/2103582
Signed-off-by: Albert Esteve <aesteve@redhat.com>