panic: Fix panic when logging.shutdown hang
We tried to add a timeout using signal.alarm() to make that vdsm will
panic if shutting down logging hangs. However this does not work, since
the alarm is delivered to the main thread instead of the thread setting
the alarm.

What actually happened after we added the alarm was:

1. Panic thread: calls signal.alarm(10)
2. Panic thread: calls logging.shutdown(), call hangs
3. Main thread: crash with "RuntimeError: Alarm timeout"

When we removed zombiereaper, we also removed the SIGALRM signal handler
by mistake, so what happens in current code is:

1. Panic thread: calls signal.alarm(10)
2. Panic thread: calls logging.shutdown(), call hangs
3. Unhandled SIGALRM terminates the process

Replace the broken alarm with a thread waiting on an event.

Fixes: 56ed0796259a (panic: Limit time spent shutting down logging)
Fixes: c0a67a0a3094 (zombiereaper: Reap out last traces of it)
Signed-off-by: Nir Soffer <nsoffer@redhat.com>