blockVolume: min sparse allocation to one extent
For block sparse volumes, if no initial_size is
defined, Vdsm allocates one chunk (2.5G), without
considering the virtual size of the volume.
This results in wasted space if the requested
volume is smaller.

Instead, we can allocate the capacity when
it is smaller than one chunk, to avoid wasting
space for small volumes.
Note than any size smaller than one extent
(128M) will be rounded to the extent size.

Bug-Url: https://bugzilla.redhat.com/2094576
Signed-off-by: Albert Esteve <aesteve@redhat.com>