virt: Add support for zero-copy migrations
Zero-copy migrations are enabled using ‘zerocopy’ migration
parameter.  It is the responsibility of Engine to pass the parameter
only when all the zero-copy migration requirements are
satisfied (parallel non-encrypted migrations).

Since adding the parameter is an API change and we don’t have a new
cluster version, a new host capability ‘zerocopy_migrations’ is
provided to indicate the presence of the zero-copy migration feature
on the host.  It is currently not used by Engine (assuming that hosts
should be kept up-to-date) but let’s have it in case we realize in
future we need it.  Note that libvirt on both the source and
destination hosts must support zero-copy migrations otherwise the
zero-copy migration fails before it starts, on an unknown libvirt
flag.

libvirt support for zero-copy migrations is not available in CentOS
Stream 8 at the moment; we do not enable zero-copy migrations there.

Bug-Url: https://bugzilla.redhat.com/2089434