spec: Use vdsm.spec directly rather than generating it
We currently generate vdsm.spec from vdsm.spec.in.  This means that
whatever we want to do regarding building Vdsm packages, we must run
auto* stuff first and having all the necessary tools available before
we start with the building process, which is not always convenient or
even possible.

The only thing that is replaced in vdsm.spec.in when generating
vdsm.spec from it is @PACKAGE_RELEASE@.  We can set 1 as its default
value and use vdsm.spec instead of vdsm.spec.in directly.  If the
value needs to be different, it can be overridden using
‘rpmbuild --define’.  This is what ‘rpm’ target in Makefile.am already
does so the release is set as intended as long as ‘make rpm’ is used
to build the rpm’s.  Tagged releases work fine with the default value.