fileVolume: split preallocate volume
When creating a new preallocated raw file volume,
the allocate_volume function first create a sparse
volume and then uses fallocate to preallocate the
bytes.

This is now used only for raw volumes,
but the same strategy can be used for cow volumes.
In order to ease using the strategy for preallocated
cow volumes, refactor the fallocate invocation
into its own function.

Signed-off-by: Albert Esteve <aesteve@redhat.com>