fileVolume: inline allocate_volume
With the refactor of _allocate_volume, the
function ended up being a wrapper for
a couple more functions that can be
invoked directly from the caller.

Also, it mixes volume creation with allocation
in a confusing way.

Instead. we can remove the _allocate_volume
function and inline everything directly from
the caller.

Signed-off-by: Albert Esteve <aesteve@redhat.com>