fileVolume: change allocate_volume exception
In fileVolume, _allocate_volume catches exceptions
from the operations module and transforms them to
VolumesZeroingError, which does not make much sense.

The operations first create an image using qemuimg,
and the preallocate the image by invoking
fallocate. In this context, a better exception to be
raised would be VolumeCreationError, which will
result in an error message like this:

Error creation a new volume: /path/to/image

Also, make sure we log the original exception, and that
we raise from it, so that the backtrace is kept.

Signed-off-by: Albert Esteve <aesteve@redhat.com>