merge_test: increase reduce vol size
Increase the size of the volumes created in the
reduce tests in the merge_test module, so that
volume sizes are more realistic.

Consequently, chunked leaf volumes do not reduce
anymore in the test. Chunked leaf volumes add an
extra chunk to the required size as the condition
to call reduce.

The test relied in the allocation
that Vdsm does when the volume is created
in `calculate_volume_alloc_size` if the create
call does not specify an initial size.
In that case, Vdsm allocates an entire chunk,
whereas the requested capacity was smaller (test
requested 2 GiB volumes, Vdsm allocated 2.5 GiB).

As long as the required size is smaller
than the difference between capacity and
chunk size (0.5 GiB), reduce will be called
during finalizeMerge.

In a real setup, tests show that leaf volumes
are not reduced during finalizeMerge except if
they have been extended previously. Thus, increasing
the size of the volumes and changing the expectation
in the test matches real scenario.

Signed-off-by: Albert Esteve <aesteve@redhat.com>