storageserver_test: add fake_connection fixture
Add fake_connection helper fixture that uses a
connection factory pattern. The ficture facilitates
a function that creates a connection with a unique
IP and ID. The factory function receives an argument
that can be used to create an invalid connection
(that will break if any test tries to connect
to it) to allow negative testing.

Signed-off-by: Albert Esteve <aesteve@redhat.com>