lvmdevices: remove try...except block
Remove try...except block around _create_system_devices.
The cmdutils.Error exception is only catched to
log a warning and raise it again. The tool will
not print the log as of now, but even if it did,
logging a warning and raising is a wrong pattern
which adds no valuable information.

Errors that happen during _create_system_devices
should be just propagated to the caller.

Signed-off-by: Albert Esteve <aesteve@redhat.com>