Fix missing informations on geo replication tab of GlusterFS volumes.
Since the upgrade to oVirt 4.5 and GlusterFS 10, information about
snapshots and geo replication are no more visible on the engine Webui.
After digging VDSM logs, I found that it's caused by the change of some
XML elements name introduced in a new Gluster release
(https://github.com/gluster/glusterfs/pull/1568) and by the addition of
the timezone in output of status command.

This commit fix the parsing process of the output returned by gluster
command, so that data sent to ovirt engine are as expected.

Bug-URL: https://bugzilla.redhat.com/show_bug.cgi?id=2104278