blocksd_test: add merge test
Add merge test in block volumes, to have
a realistic scenario with volumes and bitmaps.

When bitmaps are added to the base volume
after the top volume is create, measuring
the top volume will not consider the base
bitmaps size, even with backing enabled.

This will result in not performing an
extend and thus, failing the merge process
with `Failed to write bitmap 'xxx' to file:
No space left on device`.

This test has been used to reproduce the error
in https://bugzilla.redhat.com/2141371

Signed-off-by: Albert Esteve <aesteve@redhat.com>