volume: allow extend cow preallocated
Currently we consider the format to
determine when to skip the extend calls,
so that cow volumes are always skipped.
However, the allocation policy is
more important. For preallocated volumes
it would make more sense to extend
the volumes so that they capacity
matches their truesize, and thus
it also matches user expectations.

NOTE: commit message was updated to reflect
the actual fix better

Bug-Url: https://bugzilla.redhat.com/2170689
Signed-off-by: Albert Esteve <aesteve@redhat.com>