blocksd_test: check actual size for extend test
Current extend_volume_skipped test checks
only that the volume remains under the original
capacity. While this covers the usecase of
the volume not being extended to the new
capacity, it does not check that the volume
size has not changed.

We can make the check more precise with
little effort.

Fixes: a558d52
Signed-off-by: Albert Esteve <aesteve@redhat.com>