Fix default route with libnl3.8+
Since libnl3.8, the nl_addr_parse function does not return 'none' anymore for
'default', but it returns a 0.0.0.0/0 addr.
https://github.com/thom311/libnl/commit/0c0aee829c1d2d468d89eda75d38e7490e4de575

But since we depend on that value in vdsm, it never reports a default
gateway anymore, causing networks to be out of sync.

So next to 'none' we also check for '0.0.0.0/0' or '::/0'.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>