vdsm: get gluster volume info from any gluster peer
The function _get_gluster_volinfo query the glusterfs volume
info the the storage server, this is translated to the gluster
client adding the parameter --remote-host which limits the
query to one server, so we are converting the storage server as a
single point of failure, if it is not available, it can led to cluster
outtage. The proposed changed let the cluster cli to use any
available gluster peer.

Signed-off-by: José Enrique Gutiérrez Mazón <josgutie@redhat.com>
Signed-off-by: Albert Esteve <aesteve@redhat.com>