tox.ini: fix builds
The fix in 98970453 did not really fix the issue with setuptools and
packaging incompatibility.

The issue is that tox creates a new virtualenv with latest
setuptools/pip from pypi. But the newest setuptools version is
incompatible with the system's packaging version (20.9 on c9s).

So first thing we do is installing packaging>=22.0 into the venv to make
sure the next installs don't fail.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>