virt: handle capacity=0 in getNextVolumeSize
We saw a case where domstats returned capacity=0 just after taking a
snapshot.
This causes a ImprobableAllocationError to be triggered and permanently
keep the VM into a hanging paused state.

As writes can't be flushed to disk, the qcow2 image can become corrupted
also when the VM is stopped.

So we add some fallback if capacity=0 so the disk is at least resized,
so we don't end up with a broken VM.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>