netlink: fix crash on parsing invalid objects
libnl knows multiple object types, one of them is route/route.
This object can be passed to rtnl_route_get_dst to get the route
destination.

But we only checked for 'route/*' in `_object_input`.
This caused objects of other types (like route/neigh) to get
passed to `_route_info`, which caused a segfault in the
`nl_addr2str` call on destination as the call to
`rtnl_route_get_dst(route)` is invalid on a non 'route/route' object.

This seems to have worked since libnl commit
014c33a6e55b88630e334e25da0a23e0a62f4b4d.
Since then the fields of the rtnl_neigh has changed, causing the
segfault.

So while it did not crash in the past, it was broken since long.
We now check the full type and only pass 'route/route' objects to
`_route_info`.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>