common: fix early stop of panic
Add a join() to the thread, otherwise it happens that the main process
is already closed and the thread is still alive (because daemon = True).

This causes issues in tests, cause the returncode is 0 instead of -9, as
there was not enough time to execute the os.killpg before the main
thread ended.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>