virt: fix partial QoS removal
Since commit 50cd77006d1b5ec30759f50bc6006b18f397bd05, vdsm uses the old
bandwidth value when updating (the QoS) on the NIC.

Now lets say we have the following active on a VM NIC:
    <bandwidth>
        <inbound average="128000" peak="384000" burst="1024000" />
        <outbound average="128000" burst="1024000" peak="384000" />
    </bandwidth>

And we pass the following update from the engine:
      <bandwidth>
        <outbound average="128000" burst="1024000" peak="384000"/>
      </bandwidth>

Then it will preserve the inbound QoS, which is not something we want.

This is a leftover from pre 'libvirt XML' oVirt era.
Cause before it needed an empty 'inbound' tag to remove the inbound QoS.
See the notes in the above commit:
'It is deleted if it is in specParams but empty ({'inbound':{}}).'

But as the complete XML is now passed to vdsm, we can just rely on the
data coming from the engine, without the need to parse the old bandwidth
data.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>